Refactor: 포인트 패키지 네이밍 변경
