+++ src/main/java/supernova/whokie/point/controller/dto/PointRequest.java
@@ -1,0 +1,14 @@
+package supernova.whokie.point.controller.dto;
+
+import lombok.Builder;
+
+public class PointRequest {
+
+    @Builder
+    public record Purchase(
+            int point
+    ) {
+
+    }
+
+}
