--- src/main/java/supernova/whokie/point_record/controller/dto/PointPurchaseRequest.java
@@ -1,13 +1,0 @@
-package supernova.whokie.point_record.controller.dto;
-
-import lombok.Builder;
-
-public class PointPurchaseRequest {
-
-    @Builder
-    public record Purchase(
-            int amount
-    ) {
-
-    }
-}
